FIX: nef preview creation when option -otherimage doesn't work (ie. first attempt) [t18562]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@12383 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/preview_preprocessing.php
@@ -252,7 +252,11 @@ if (($extension=="cr2" || $extension=="nef" || $extension=="dng" || $extension==
             // check for nef -otherimage failure
             if ($extension=="nef"&&!filesize_unlimited($target)>0)
                 {
-                unlink($target);    
+                if(file_exists($target))
+                    {
+                    unlink($target);
+                    }
+
                 $bin_tag=" -previewimage ";
                 //2nd attempt
                 $cmd=$exiftool_fullpath.' -b '.$bin_tag.' '.escapeshellarg($file).' -w %d%f.jpg';
